Secure integration web service calls HAJ-340

+++ shj-admin-portal-orm/src/main/java/com/elm/shj/admin/portal/orm/entity/JpaAuditLog.java
@@ -31,7 +31,7 @@ public class JpaAuditLog implements Serializable {
     private long id;
 
     @Column(name = "user_id_number", nullable = false)
-    private int userIdNumber;
+    private long userIdNumber;
 
     @Column(nullable = false)
     private String handler;
